pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nazitest"
version = "0.1.0"
description = "AI-powered autonomous penetration testing framework"
requires-python = ">=3.11"
dependencies = [
    # Browser automation
    "zendriver",
    "pydoll-python",
    # TLS-safe HTTP
    "curl_cffi",
    # Knowledge graph
    "networkx",
    # Code analysis
    "tree-sitter",
    "tree-sitter-language-pack",
    # CLI
    "typer[all]",
    "rich",
    # Data
    "orjson",
    "haralyzer",
    "beautifulsoup4",
    "lxml",
    # Async
    "uvloop",
    # Reporting
    "jinja2",
    "weasyprint",
    # LLM
    "openai",
    "httpx",
    "tiktoken",
    "pydantic>=2.0",
    # Security & scope
    "tldextract",
    "publicsuffixlist",
    "cryptography",
    # YAML config
    "pyyaml",
]

[project.scripts]
nazitest = "nazitest.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
